Move ApiError into the error module
