+++ src/lz-web/src/api.rs
@@ -5,11 +5,9 @@
 mod observability;
 mod searching;
 
-use std::fmt;
 use std::sync::Arc;
 
 use axum::extract::Query;
-use axum::http::StatusCode;
 use axum::routing::{get, post};
 use axum::{debug_handler, Json, Router};
 use lz_db::{
@@ -25,7 +23,10 @@ use utoipa::{IntoParams, OpenApi, ToResponse, ToSchema};
 
 use crate::db::queries::{list_bookmarks, AnnotatedBookmark, ListResult, Pagination};
 use crate::db::{DbTransaction, GlobalWebAppState};
-use crate::http::{lookup_page_from_web, LookupError, Metadata};
+use crate::http::{lookup_page_from_web, Metadata};
+
+mod error;
+use error::ApiError;
 
 #[derive(OpenApi)]
 #[openapi(
@@ -48,78 +49,6 @@ pub fn router() -> Router<Arc<GlobalWebAppState>> {
         .route("/tag/complete", get(complete_tag))
         .layer(CorsLayer::permissive());
     observability::add_layers(router)
-}
-
-#[derive(Serialize, Deserialize, Debug, ToSchema)]
-enum ApiError {
-    #[schema()]
-    NotFound,
-
-    #[schema()]
-    #[serde(serialize_with = "serialize_db_error", skip_deserializing)]
-    DatastoreError(sqlx::Error),
-
-    #[schema()]
-    #[serde(serialize_with = "serialize_lookup_error", skip_deserializing)]
-    Lookup(LookupError),
-}
-
-impl fmt::Display for ApiError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            ApiError::NotFound => write!(f, "not found"),
-            ApiError::DatastoreError(_) => write!(f, "datastore error"),
-            ApiError::Lookup(e) => write!(f, "HTTP error {e}"),
-        }
-    }
-}
-
-impl axum::response::IntoResponse for ApiError {
-    fn into_response(self) -> axum::response::Response {
-        #[derive(Serialize)]
-        struct ErrorResponse<'a> {
-            error_message: &'a str,
-        }
-
-        let error_as_text = self.to_string();
-        let (status, error_message) = match &self {
-            ApiError::NotFound => (StatusCode::NOT_FOUND, &error_as_text),
-            ApiError::DatastoreError(inner) => {
-                tracing::error!(error=%inner, error_debug=?inner, "datastore error");
-                (StatusCode::INTERNAL_SERVER_ERROR, &error_as_text)
-            }
-            ApiError::Lookup(error) => {
-                tracing::warn!(error_debug=?error, %error, "HTTP error");
-                (StatusCode::BAD_REQUEST, &error_as_text)
-            }
-        };
-        (status, Json(ErrorResponse { error_message })).into_response()
-    }
-}
-
-impl From<sqlx::Error> for ApiError {
-    fn from(other: sqlx::Error) -> Self {
-        ApiError::DatastoreError(other)
-    }
-}
-impl From<LookupError> for ApiError {
-    fn from(other: LookupError) -> Self {
-        ApiError::Lookup(other)
-    }
-}
-
-fn serialize_db_error<S>(_err: &sqlx::Error, s: S) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    s.serialize_str("(nasty DB error omitted)")
-}
-
-fn serialize_lookup_error<S>(err: &LookupError, s: S) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    s.serialize_str(&format!("lookup error: {}", err))
 }
 
 /// The response returned by the `list_bookmarks` API endpoint.
